+++ lib/doorkeeper/orm/active_record/application.rb
@@ -70,10 +70,27 @@ module Doorkeeper
       end
     end
 
-    def to_json(options = nil)
-      serializable_hash(except: :secret)
-        .merge(secret: plaintext_secret)
-        .to_json(options)
+    # Represents client as set of it's attributes in JSON format.
+    # This is the right way how we want to override ActiveRecord #to_json.
+    #
+    # Respects privacy settings and serializes minimum set of attributes
+    # for public/private clients and full set for authorized owners.
+    #
+    # @return [Hash] entity attributes for JSON
+    #
+    def as_json(options = {})
+      # if application belongs to some owner we need to check if it's the same as
+      # the one passed in the options or check if we render the client as an owner
+      if (respond_to?(:owner) && owner && owner == options[:current_resource_owner]) ||
+         options[:as_owner]
+        # Owners can see all the client attributes, fallback to ActiveModel serialization
+        super
+      else
+        # if application has no owner or it's owner doesn't match one from the options
+        # we render only minimum set of attributes that could be exposed to a public
+        only = extract_serializable_attributes(options)
+        super(options.merge(only: only))
+      end
     end
 
     private
@@ -97,6 +114,49 @@ module Doorkeeper
 
     def enforce_scopes?
       Doorkeeper.configuration.enforce_configured_scopes?
+    end
+
+    # Helper method to extract collection of serializable attribute names
+    # considering serialization options (like `only`, `except` and so on).
+    #
+    # @param options [Hash] serialization options
+    #
+    # @return [Array<String>]
+    #   collection of attributes to be serialized using #as_json
+    #
+    def extract_serializable_attributes(options = {})
+      opts = options.try(:dup) || {}
+      only = Array.wrap(opts[:only]).map(&:to_s)
+
+      only = if only.blank?
+               serializable_attributes
+             else
+               only & serializable_attributes
+             end
+
+      only -= Array.wrap(opts[:except]).map(&:to_s) if opts.key?(:except)
+      only.uniq
+    end
+
+    # We need to hook into this method to allow serializing plan-text secrets
+    # when secrets hashing enabled.
+    #
+    # @param key [String] attribute name
+    #
+    def read_attribute_for_serialization(key)
+      return super unless key.to_s == "secret"
+
+      plaintext_secret || secret
+    end
+
+    # Collection of attributes that could be serialized for public.
+    # Override this method if you need additional attributes to be serialized.
+    #
+    # @return [Array<String>] collection of serializable attributes
+    def serializable_attributes
+      attributes = %w[id name created_at]
+      attributes << "uid" unless confidential?
+      attributes
     end
   end
 end